move QueueManager to separate module
